hub/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
websockets==12.0
requests==2.31.0
cryptography==41.0.0
pandas==2.1.0
numpy==1.26.0
pyjwt==2.8.0
aiohttp==3.9.1
